No longer need pre-processing script to combine rotation files.
As of pygplates revision 25, pygplates.RotationModel is much faster, so we don't need to reduce the set of rotations, in rotation files, to only those needed. Also determining which rotation plate IDs are needed is will be more difficult when plate velocities are added (and used in deforming regions), so it's best not to rely on this.